--- a/integration/cypress/integration/subnets/subnets.spec.ts
+++ b/integration/cypress/integration/subnets/subnets.spec.ts
@@ -46,7 +46,7 @@ context("Subnets", () => {
     cy.visit(generateNewURL("/networks"));
 
     cy.findByRole("combobox", { name: "Group by" }).within(() => {
-      cy.findByRole("option", { selected: true }).contains("Fabric");
+      cy.findByRole("option", { selected: true }).contains("Group by fabric");
     });
 
     cy.url().should("include", generateNewURL("/networks?by=fabric"));
@@ -58,11 +58,11 @@ context("Subnets", () => {
     });
 
     cy.findByRole("combobox", { name: "Group by" }).within(() => {
-      cy.findByRole("option", { selected: true }).contains("Fabric");
-      cy.findByRole("option", { selected: false }).contains("Space");
+      cy.findByRole("option", { selected: true }).contains("Group by fabric");
+      cy.findByRole("option", { selected: false }).contains("Group by space");
     });
 
-    cy.findByRole("combobox", { name: "Group by" }).select("Space");
+    cy.findByRole("combobox", { name: "Group by" }).select("Group by space");
 
     cy.findByRole("table", { name: "Subnets by Space" }).within(() => {
       cy.findAllByRole("columnheader").first().should("have.text", "Space");