--- a/ui/src/app/subnets/views/SubnetsList/SubnetsTable/SubnetsTable.tsx
+++ b/ui/src/app/subnets/views/SubnetsList/SubnetsTable/SubnetsTable.tsx
@@ -1,6 +1,8 @@
+import { useState } from "react";
+
 import FabricTable from "./FabricTable";
 import SpaceTable from "./SpaceTable";
-import { useSubnetsTable } from "./hooks";
+import { useSubnetsTable, useSubnetsTableSearch } from "./hooks";
 
 import SubnetsControls from "app/subnets/views/SubnetsList/SubnetsControls";
 import type { SubnetGroupByProps } from "app/subnets/views/SubnetsList/SubnetsTable/types";
@@ -9,16 +11,28 @@ const SubnetsTable = ({
   groupBy,
   setGroupBy,
 }: SubnetGroupByProps): JSX.Element | null => {
-  const { rows } = useSubnetsTable(groupBy);
+  const [searchText, setSearchText] = useState("");
+  const { rows, loaded } = useSubnetsTable(groupBy);
+  const filteredRows = useSubnetsTableSearch(rows, searchText);
+
+  const handleSearch = (searchText: string): void => {
+    setSearchText(searchText);
+  };
+
+  const emptyMsg = !loaded ? "Loading..." : "No results found";
 
   return (
     <>
-      <SubnetsControls groupBy={groupBy} setGroupBy={setGroupBy} />
+      <SubnetsControls
+        groupBy={groupBy}
+        setGroupBy={setGroupBy}
+        handleSearch={handleSearch}
+      />
 
       {groupBy === "fabric" ? (
-        <FabricTable data={rows} />
+        <FabricTable data={filteredRows} emptyMsg={emptyMsg} />
       ) : (
-        <SpaceTable data={rows} />
+        <SpaceTable data={filteredRows} emptyMsg={emptyMsg} />
       )}
     </>
   );