--- a/ui/src/app/subnets/views/SubnetsList/SubnetsTable/constants.ts
+++ b/ui/src/app/subnets/views/SubnetsList/SubnetsTable/constants.ts
@@ -7,13 +7,13 @@ export enum SubnetsColumns {
   SPACE = "space",
 }
 
-export const subnetColumnLabels: Record<SubnetsColumns, string> = {
+export const subnetColumnLabels = {
   [SubnetsColumns.FABRIC]: "Fabric",
   [SubnetsColumns.VLAN]: "VLAN",
   [SubnetsColumns.DHCP]: "DHCP",
   [SubnetsColumns.SUBNET]: "Subnet",
   [SubnetsColumns.IPS]: "Available IPs",
   [SubnetsColumns.SPACE]: "Space",
-};
+} as const;
 
 export const SUBNETS_TABLE_ITEMS_PER_PAGE = 25;