--- a/ui/src/app/subnets/views/SubnetsList/_index.scss
+++ b/ui/src/app/subnets/views/SubnetsList/_index.scss
@@ -55,4 +55,9 @@
       width: 1px;
     }
   }
+
+  .subnets-controls-group-by {
+    display: flex;
+    align-items: baseline;
+  }
 }