@petermakowski 2022-03-23
> Wow, this is great, and a _huge_ improvement over the original legacy version! I think there are some minor improvements that could be made (e.g. adding search query to URL, "No results found" flashes on the screen, maybe some help text explaining what the search field searches by) but for sure this can land as-is. 👍

Yup, these are all good suggestions, will look into next. Thanks.